FIxed bug that prevented mips from sending
Fixed the bug that prevented mips from being sent. Needed to change the datatype of the mips.

+++ mipTesting.py
@@ -55,7 +55,7 @@ def mipTesting(mask,flow,mag):
 
 
 
-    mipimage = mipimage.astype(int)
+    mipimage = mipimage.astype(np.int16)
     io.savemat('thing.mat',{'thing':mipimage})
     return mipimage
 """
